requirements.txt
altgraph==0.17.4
et-xmlfile==1.1.0
lxml==5.3.0
macholib==1.16.3
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
xlrd==2.0.1
